Code tidy a wider set of code.

+++ lang/src/test/java/net/openhft/lang/io/serialization/JavaSerializationTest.java
@@ -33,7 +33,7 @@ public class JavaSerializationTest {
     @Test
     @Ignore
     public void NullPointerException() {
-        final Bytes bytes = ByteBufferBytes.wrap(ByteBuffer.allocate((int) 1024).order
+        final Bytes bytes = ByteBufferBytes.wrap(ByteBuffer.allocate(1024).order
                 (ByteOrder.nativeOrder()));
 
         NullPointerException expected = new NullPointerException("test");
